Fix "block" filling lines with single lines to 40 chars
The "block" justifier, whose job is to place spaces between words as to always fill a line from end to end, fucked up when encountering lines with only one token (word or symbol - no spaces) and would repeat it until the line was 40 characters long.

This was fixed, and parts of the "pandora" page which depended on the bug being present was reconstructed explicitly.