Reintroduce 1D cellular automata
Reintroduces the cellular automata curtains (which was present in the PHP version). It's a very simple thing which i reccomend reading up on. I tink theres a 1/3rd chance of the cellular automata getting chosen as curtains, and then it picks a random algorithm from a list of ones i like.

Per definition there's only 256 ones avaliable in this scheme. I think i picked seven or so.